[Feat] #35 - 로그인 후 accessToken, accessTokenExpireIn DataStore에 저장

+++ data/src/main/java/com/yapp/buddycon/data/network/response/UserInfoResponse.kt
@@ -3,5 +3,5 @@ package com.yapp.buddycon.data.network.response
 // TODO(OWS) : 필드 값 변경
 data class UserInfoResponse(
     val accessToken: String,
-    val accessTokenExpiresIn: Int
+    val accessTokenExpiresIn: Long
 )
