+++ domain/src/main/java/com/yapp/buddycon/domain/model/UserInfo.kt
@@ -3,5 +3,5 @@ package com.yapp.buddycon.domain.model
 // TODO(OWS) : 필드 값 변경
 data class UserInfo(
     val accessToken: String,
-    val accessTokenExpiresIn: Int
+    val accessTokenExpiresIn: Long
 )
